fix search error
